src/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.11
aiosignal==1.3.1
astroid==3.0.3
attrs==24.2.0
aws-lambda-powertools==2.27.1
aws-xray-sdk==2.14.0
black==24.3.0
boto3==1.34.136
botocore==1.34.158
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
cryptography==44.0.0
dill==0.3.8
distlib==0.3.8
execnet==2.1.1
fastjsonschema==2.20.0
filelock==3.15.4
frozenlist==1.4.1
identify==2.6.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.5
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mccabe==0.7.0
moto==5.0.11
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.6.0
pyclean==3.0.0
pycparser==2.22
pylint==3.0.3
pytest==8.1.1
pytest-cov==5.0.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
responses==0.25.3
rpds-py==0.20.0
s3transfer==0.10.2
shellcheck-py==0.10.0.1
six==1.16.0
tomlkit==0.13.0
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
Werkzeug==3.0.6
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.9.4
